ml/requirements.txt
scikit-learn>=1.3
shap>=0.43
pandas>=2.0
numpy>=1.24
joblib>=1.3
pyyaml>=6.0
matplotlib>=3.7
streamlit>=1.30
